added progress wheel to PhotoMediaFragment
