+++ app/src/main/java/com/freshdigitable/udonroad/media/PhotoMediaFragment.java
@@ -16,25 +16,17 @@
 
 package com.freshdigitable.udonroad.media;
 
-import android.content.Context;
-import android.graphics.Canvas;
-import android.graphics.Matrix;
-import android.graphics.RectF;
-import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v4.view.GestureDetectorCompat;
-import android.support.v7.widget.AppCompatImageView;
-import android.view.GestureDetector.SimpleOnGestureListener;
 import android.view.LayoutInflater;
-import android.view.MotionEvent;
-import android.view.ScaleGestureDetector;
-import android.view.ScaleGestureDetector.OnScaleGestureListener;
-import android.view.ScaleGestureDetector.SimpleOnScaleGestureListener;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
+import android.widget.ProgressBar;
+import android.widget.Toast;
 
+import com.freshdigitable.udonroad.R;
+import com.squareup.picasso.Callback;
 import com.squareup.picasso.Picasso;
 
 /**
@@ -46,6 +38,7 @@ public class PhotoMediaFragment extends MediaViewActivity.MediaFragment {
   @SuppressWarnings("unused")
   private static final String TAG = PhotoMediaFragment.class.getSimpleName();
   private ImageView imageView;
+  private ProgressBar progressBar;
   private String loadingTag;
 
   @Nullable
@@ -53,8 +46,10 @@ public class PhotoMediaFragment extends MediaViewActivity.MediaFragment {
   public View onCreateView(LayoutInflater inflater,
                            @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
     super.onCreateView(inflater, container, savedInstanceState);
-    imageView = new ScalableImageView(getContext());
-    return imageView;
+    final View v = inflater.inflate(R.layout.fragment_photo_media, container, false);
+    imageView = v.findViewById(R.id.media_image_view);
+    progressBar = v.findViewById(R.id.media_progress_bar);
+    return v;
   }
 
   @Override
@@ -66,7 +61,17 @@ public class PhotoMediaFragment extends MediaViewActivity.MediaFragment {
     Picasso.with(getContext())
         .load(mediaEntity.getMediaURLHttps() + ":medium")
         .tag(loadingTag)
-        .into(imageView);
+        .into(imageView, new Callback() {
+          @Override
+          public void onSuccess() {
+            progressBar.setVisibility(View.GONE);
+          }
+
+          @Override
+          public void onError() {
+            Toast.makeText(getContext(), "failed to load media...", Toast.LENGTH_LONG).show();
+          }
+        });
   }
 
   @Override
@@ -76,159 +81,5 @@ public class PhotoMediaFragment extends MediaViewActivity.MediaFragment {
     imageView.setOnClickListener(null);
     imageView.setOnTouchListener(null);
     imageView.setImageDrawable(null);
-  }
-
-  private static class ScalableImageView extends AppCompatImageView {
-    private final GestureDetectorCompat gestureDetector;
-    private final ScaleGestureDetector scaleGestureDetector;
-
-    public ScalableImageView(Context context) {
-      super(context, null, 0);
-      scaleGestureDetector = new ScaleGestureDetector(getContext(), scaleListener);
-      gestureDetector = new GestureDetectorCompat(getContext(), scrollListener);
-      gestureDetector.setIsLongpressEnabled(false);
-      setScaleType(ScaleType.MATRIX);
-    }
-
-    @Override
-    public boolean onTouchEvent(MotionEvent event) {
-      scaleGestureDetector.onTouchEvent(event);
-      final boolean scaling = scaleGestureDetector.isInProgress();
-      if (scaling) {
-        transformMatrix.postScale(scale, scale, focusX, focusY);
-      }
-      final boolean scrolled = gestureDetector.onTouchEvent(event);
-      if (scrolled) {
-        transformMatrix.postTranslate(transX, transY);
-      }
-      final boolean invalidated = scaling || scrolled;
-      if (invalidated) {
-        getParent().requestDisallowInterceptTouchEvent(true);
-        invalidate();
-      }
-      return invalidated || super.onTouchEvent(event);
-    }
-
-    private final float[] imageMat = new float[9];
-
-    @Override
-    protected void onDraw(Canvas canvas) {
-      if (!transformMatrix.isIdentity()) {
-        final Matrix matrix = getImageMatrix();
-        matrix.postConcat(transformMatrix);
-        matrix.getValues(imageMat);
-        imageMat[Matrix.MSCALE_X] = Math.max(imageMat[Matrix.MSCALE_X], matToFit[Matrix.MSCALE_X]);
-        imageMat[Matrix.MSCALE_Y] = Math.max(imageMat[Matrix.MSCALE_Y], matToFit[Matrix.MSCALE_Y]);
-
-        final float maxTransX = getWidth() - getDrawable().getIntrinsicWidth() * imageMat[Matrix.MSCALE_X];
-        if (Math.abs(maxTransX) < Math.abs(imageMat[Matrix.MTRANS_X])) {
-          imageMat[Matrix.MTRANS_X] = maxTransX;
-        } else if (Math.signum(maxTransX) * imageMat[Matrix.MTRANS_X] < 0) {
-          imageMat[Matrix.MTRANS_X] = 0;
-        }
-
-        final float maxTransY = getHeight() - getDrawable().getIntrinsicHeight() * imageMat[Matrix.MSCALE_Y];
-        if (Math.abs(maxTransY) < Math.abs(imageMat[Matrix.MTRANS_Y])) {
-          imageMat[Matrix.MTRANS_Y] = maxTransY;
-        } else if (Math.signum(maxTransY) * imageMat[Matrix.MTRANS_Y] < 0) {
-          imageMat[Matrix.MTRANS_Y] = 0;
-        }
-
-        matrix.setValues(imageMat);
-        setImageMatrix(matrix);
-        transformMatrix.reset();
-      }
-      super.onDraw(canvas);
-    }
-
-    private Matrix transformMatrix = new Matrix();
-    private float scale = 1;
-    private float focusX;
-    private float focusY;
-    private float transX;
-    private float transY;
-
-    private final SimpleOnGestureListener scrollListener = new SimpleOnGestureListener() {
-      @Override
-      public boolean onScroll(MotionEvent e1, MotionEvent e2, float distanceX, float distanceY) {
-        if (e2.getEventTime() - e1.getEventTime() < 200
-            && shouldGoNextPage(distanceX)) {
-          return false;
-        }
-        transX = -distanceX;
-        transY = -distanceY;
-        return true;
-      }
-
-      private boolean shouldGoNextPage(final float distX) {
-        if (Math.abs(distX) < 0.1) {
-          return false;
-        }
-        if (getDrawable() == null) {
-          return true;
-        }
-        if (getHeight() == getDrawable().getIntrinsicHeight() * imageMat[Matrix.MSCALE_Y]) {
-          return true;
-        }
-        if (distX > 0) {
-          final float maxTransX = getWidth() - getDrawable().getIntrinsicWidth() * imageMat[Matrix.MSCALE_X];
-          return maxTransX <= 0
-              && Math.abs(maxTransX - imageMat[Matrix.MTRANS_X]) < 0.01;
-        } else {
-          return Math.abs(imageMat[Matrix.MTRANS_X]) < 0.01;
-        }
-      }
-    };
-
-    private final OnScaleGestureListener scaleListener = new SimpleOnScaleGestureListener() {
-      @Override
-      public boolean onScale(ScaleGestureDetector detector) {
-        scale = detector.getScaleFactor();
-        focusX = detector.getFocusX();
-        focusY = detector.getFocusY();
-        return true;
-      }
-    };
-
-    private final RectF viewRect = new RectF();
-    private final RectF drawableRect = new RectF();
-    private final Matrix matrixToFit = new Matrix();
-    private final float[] matToFit = new float[9];
-
-    @Override
-    protected void onSizeChanged(int w, int h, int oldw, int oldh) {
-      super.onSizeChanged(w, h, oldw, oldh);
-      viewRect.set(0, 0, w, h);
-      updateMatrixToFit();
-    }
-
-    @Override
-    public void setImageDrawable(@Nullable Drawable drawable) {
-      super.setImageDrawable(drawable);
-      if (drawable == null) {
-        drawableRect.setEmpty();
-      } else {
-        drawableRect.set(0, 0, drawable.getIntrinsicWidth(), drawable.getIntrinsicHeight());
-      }
-      updateMatrixToFit();
-    }
-
-    private void updateMatrixToFit() {
-      if (drawableRect.isEmpty() || viewRect.isEmpty()) {
-        matrixToFit.reset();
-        transformMatrix.reset();
-      } else {
-        matrixToFit.setRectToRect(drawableRect, viewRect, Matrix.ScaleToFit.CENTER);
-        setImageMatrix(matrixToFit);
-        invalidate();
-      }
-      matrixToFit.getValues(matToFit);
-    }
-
-    @Override
-    public void setImageMatrix(Matrix matrix) {
-      super.setImageMatrix(matrix);
-      matrix.getValues(imageMat);
-    }
   }
 }
